Refactor PysimSimulator
