+++ transactron/testing/__init__.py
@@ -1,7 +1,8 @@
 from amaranth.sim._async import TestbenchContext, ProcessContext, SimulatorContext  # noqa: F401
 from .input_generation import *  # noqa: F401
-from .functions import *  # noqa: F401
 from .infrastructure import *  # noqa: F401
+from .functions import *  # noqa: F401
+from .simulator import *  # noqa: F401
 from .test_circuit import *  # noqa: F401
 from .method_mock import *  # noqa: F401
 from .testbenchio import *  # noqa: F401
